+++ core/dbt/artifacts/schemas/run/v5/run.py
@@ -4,7 +4,7 @@ import copy
 import threading
 from dataclasses import dataclass, field
 from datetime import datetime
-from typing import Any, Dict, Iterable, List, Optional, Sequence, Tuple
+from typing import Any, Dict, Iterable, Optional, Sequence, Tuple
 
 # https://github.com/dbt-labs/dbt-core/issues/10098
 # Needed for Mashumaro serialization of RunResult below
@@ -19,6 +19,7 @@ from dbt.artifacts.schemas.base import (
     get_artifact_schema_version,
     schema_version,
 )
+from dbt.artifacts.schemas.batch_results import BatchResults
 from dbt.artifacts.schemas.results import (
     BaseResult,
     ExecutionResult,
@@ -29,21 +30,6 @@ from dbt.artifacts.schemas.results import (
 from dbt.exceptions import scrub_secrets
 from dbt_common.clients.system import write_json
 from dbt_common.constants import SECRET_ENV_PREFIX
-from dbt_common.dataclass_schema import dbtClassMixin
-
-BatchType = Tuple[Optional[datetime], datetime]
-
-
-@dataclass
-class BatchResults(dbtClassMixin):
-    successful: List[BatchType] = field(default_factory=list)
-    failed: List[BatchType] = field(default_factory=list)
-
-    def __add__(self, other: BatchResults) -> BatchResults:
-        return BatchResults(
-            successful=self.successful + other.successful,
-            failed=self.failed + other.failed,
-        )
 
 
 @dataclass
