+++ tests/unit/task/test_run.py
@@ -6,8 +6,9 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from dbt.adapters.postgres import PostgresAdapter
+from dbt.artifacts.schemas.batch_results import BatchResults
 from dbt.artifacts.schemas.results import RunStatus
-from dbt.artifacts.schemas.run import BatchResults, RunResult
+from dbt.artifacts.schemas.run import RunResult
 from dbt.config.runtime import RuntimeConfig
 from dbt.contracts.graph.manifest import Manifest
 from dbt.contracts.graph.nodes import ModelNode
